+++ xskillscore/tests/test_probabilistic.py
@@ -14,7 +14,6 @@ from xskillscore.core.probabilistic import (
     crps_gaussian,
     crps_quadrature,
     discrimination,
-    fair_brier_score,
     rank_histogram,
     reliability,
     rps,
@@ -248,12 +247,14 @@ def test_threshold_brier_score_multiple_thresholds_dask(
         assert actual.attrs == {}
 
 
+@pytest.mark.parametrize("fair_bool", [True, False])
 @pytest.mark.parametrize("keep_attrs", [True, False])
-def test_brier_score(o, f_prob, keep_attrs):
+def test_brier_score(o, f_prob, keep_attrs, fair_bool):
     actual = brier_score(
         (o > 0.5).assign_attrs(**o.attrs),
         (f_prob > 0.5).mean("member"),
         keep_attrs=keep_attrs,
+        fair=fair_bool
     )
     assert actual.chunks is None or actual.chunks == ()
     if keep_attrs:
@@ -261,43 +262,17 @@ def test_brier_score(o, f_prob, keep_attrs):
     else:
         assert actual.attrs == {}
 
-
-@pytest.mark.parametrize("keep_attrs", [True, False])
-def test_fair_brier_score(o, f_prob, keep_attrs):
-    actual = fair_brier_score(
-        (o > 0.5).assign_attrs(**o.attrs),
-        (f_prob > 0.5),
-        keep_attrs=keep_attrs,
-    )
-    assert actual.chunks is None or actual.chunks == ()
-    if keep_attrs:
-        assert actual.attrs == o.attrs
-    else:
-        assert actual.attrs == {}
 
 
 @pytest.mark.parametrize("dim", DIMS)
-def test_fair_brier_score_dim(o, f_prob, dim):
-    actual = fair_brier_score((o > 0.5), (f_prob > 0.5), dim=dim)
-    assert_only_dim_reduced(dim, actual, o)
+def test_brier_score_vs_fair_brier_score(o, f_prob,dim):
+    """Test that brier_score scores lower for limited ensemble than bias brier_score."""
+    fbs = brier_score((o > 0.5), (f_prob > 0.5), dim=dim, fair=True)
+    bs = brier_score((o > 0.5), (f_prob > 0.5).mean("member"), dim=dim, fair=False)
+    assert (fbs <= bs).all(),print("fairBS", fbs, "\nBS", bs)
+    #assert (fbs <= 1).all()
+    #assert (fbs >= 0).all()
 
-
-def test_brier_score_vs_fair_brier_score(o, f_prob):
-    dim = ["lon", "lat"]
-    fbs = fair_brier_score((o > 0.5), (f_prob > 0.5), dim=dim)
-    bs = brier_score((o > 0.5), (f_prob > 0.5).mean("member"), dim=dim)
-    print("fairBS", fbs, "\nBS", bs)
-    assert (fbs <= bs).all()
-    assert (fbs <= 1).all()
-    assert (fbs >= 0).all()
-
-
-def test_fair_brier_score_dask(o_dask, f_prob_dask):
-    actual = brier_score(
-        (o_dask > 0.5).assign_attrs(**o_dask.attrs),
-        (f_prob_dask > 0.5),
-    )
-    assert is_dask_collection(actual)
 
 
 @pytest.mark.parametrize("dim", DIMS)
@@ -314,18 +289,21 @@ def test_threshold_brier_score_dask_threshold(o_dask, f_prob_dask, threshold):
     assert actual.chunks is not None
 
 
+@pytest.mark.parametrize("fair_bool", [True, False])
 @pytest.mark.parametrize("dim", DIMS)
-def test_brier_score_dim(o, f_prob, dim):
-    actual = brier_score((o > 0.5), (f_prob > 0.5).mean("member"), dim=dim)
+def test_brier_score_dim(o, f_prob, dim,fair_bool):
+    actual = brier_score((o > 0.5), (f_prob > 0.5).mean("member"), dim=dim,fair=fair_bool)
     assert_only_dim_reduced(dim, actual, o)
 
 
+@pytest.mark.parametrize("fair_bool", [True, False])
 @pytest.mark.parametrize("keep_attrs", [True, False])
-def test_brier_score_dask(o_dask, f_prob_dask, keep_attrs):
+def test_brier_score_dask(o_dask, f_prob_dask, keep_attrs,fair_bool):
     actual = brier_score(
         (o_dask > 0.5).assign_attrs(**o_dask.attrs),
         (f_prob_dask > 0.5).mean("member"),
         keep_attrs=keep_attrs,
+        fair=fair_bool
     )
     assert actual.chunks is not None
     expected = properscoring.brier_score(
@@ -475,11 +453,12 @@ def test_reliability_dask(o_dask, f_prob_dask):
     assert actual.chunks is not None
 
 
+@pytest.mark.parametrize("fair_bool", [True, False])
 @pytest.mark.parametrize("dim", DIMS)
 @pytest.mark.parametrize("obj", ["da", "ds", "chunked_da", "chunked_ds"])
-def test_rps(o, f_prob, category_edges, dim, obj):
+def test_rps_reduce_dim(o, f_prob, category_edges, dim, obj,fair_bool):
     """Test that rps reduced dim and works for (chunked) ds and da"""
-    actual = rps(o, f_prob, category_edges=category_edges, dim=dim)
+    actual = rps(o, f_prob, category_edges=category_edges, dim=dim,fair=fair_bool)
     assert_only_dim_reduced(dim, actual, o)
 
 
@@ -504,13 +483,31 @@ def test_2_category_rps_equals_brier_score(o, f_prob):
     )
 
 
-def test_rps_perfect_values(o, category_edges):
+@pytest.mark.parametrize("fair_bool", [True, False])
+def test_rps_perfect_values(o, category_edges,fair_bool):
     """Test values for perfect forecast"""
     f = xr.concat(10 * [o], dim="member")
-    res = rps(o, f, category_edges=category_edges)
+    res = rps(o, f, category_edges=category_edges,fair=fair_bool)
     assert (res == 0).all()
 
 
-def test_rps_dask(o_dask, f_prob_dask, category_edges):
+#@pytest.mark.skip(reason='np.clip needs to be fixed')
+@pytest.mark.parametrize("fair_bool", [True, False])
+def test_rps_dask(o_dask, f_prob_dask, category_edges,fair_bool):
     """Test that rps returns dask array if provided dask array"""
-    assert rps(o_dask, f_prob_dask, category_edges=category_edges).chunks is not None
+    assert rps(o_dask, f_prob_dask, category_edges=category_edges,fair=fair_bool).chunks is not None
+
+
+@pytest.mark.parametrize("dim", DIMS)
+def test_rps_vs_fair_rps(o, f_prob, category_edges,dim):
+    frps = rps(o, f_prob, dim=dim,fair=True,category_edges=category_edges)
+    ufrps = rps(o, f_prob, dim=dim, fair=False,category_edges=category_edges)
+    assert (frps <= ufrps).all(), print("fairrps", frps, "\nufrps", ufrps)
+
+
+@pytest.mark.parametrize("dim", DIMS)
+@pytest.mark.parametrize("fair_bool", [True, False])
+def test_rps_limits(o, f_prob, category_edges, fair_bool,dim):
+    res = rps(o, f_prob, dim=dim,fair=fair_bool,category_edges=category_edges)
+    assert (res <= 1.).all(), print(res.max())
+    assert (res >= 0).all(), print(res.min())